fix broken paths issue for nav links
The site suffered through some broken link issues because
during production builds gatsbyjs does server-side rendering.
The new approach is to pass the config to all the components
as a prop. Each component has functionality that they don't
necessarily need. This is important because we shouldn't refactor
this bug into production again in the future.

closes #26

summary of changes:

Fix /null on nav panel close button. The button for some reason
was causing a null error when the button was clicked. Not sure
why but it seems gatsbyjs does a lot of behind the scenes monitering
on anchor tags. Best to avoid using them unless deemed necessary.

The nav links do not prevent undefined.

The click handlers are working again.

Update component props to have the proper config variables.

Update the test suite to check for config props.